@@ -1,11 +1,17 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import { BrowserRouter, Routes, Route, } from 'react-router-dom';
+import SinglePage from './components/page/SinglePage';
 
-ReactDOM.render(
-  <>
-    <App />
-  </>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(
+  <BrowserRouter>
+    <Routes>
+      <Route path="/" element={<App />} />
+        <Route path="/:id" element={<SinglePage />} />
+    </Routes>
+  </BrowserRouter>
 );
