@@ -1,8 +1,8 @@
 import React from 'react'
 
-const Button = ( { name }) => {
+const Button = ( { name, mode }) => {
   return (
-    <div className='text-white'>
+    <div className={mode === true ? 'text-white cursor-pointer' : "text-black cursor-pointer"}>
       {name}
     </div>
   )
